Use new DataTypes module in angular front end  (#80)
* angular front end to talk to api with new DataTransferRequest type.
Angular XHR doesnt follow redirects from POST without an intial
preflight task. To get around this, we dont redirect from configure and
instead return a url in the response to follow. (Similar to the way we
display a page for import setup and redirect to the auth flow).
* rename /configure endpoint to /_/configure
* Rename DataTransfer type to DataTransferResponse
* Rename ConfigureHandler to DataTransferHandler
* Use interface for DataTransferResponse instead of class on angular
frontend. This is so that the response of the POST request isnt bound to
a type but rather to the interface.
* formatting fixes, new lines, license headers

This requires a new build and static content deployment (tracked in #67)