Remove jobId from PortabilityJob (both LegacyPortabilityJob and the new PortabilityJob via EntityType)
It is redundant to store it in the job as we are keying off of it in our stores.

Also change jobID to be type UUID instead of String, for a few reasons:
- UUIDs are used directly in CosmosStore
- Prevent confusion around whether String jobIDs passed around are encoded are not
- All impls are using UUID anyway and converting to and from String

I originally thought this would be necessary for #132, turns out it's not but is still good to do.