Add job lifetime hooks interface
This introduces some changes as to how the end of a job is handled. Current behaviour is to simply return when the data copier fails because of a RetryException for example. We introduce the CopyException to throw here instead. We also add a new JobHooks interface with two  methods that are called at the start and end of a job. A cloud extension can then provide an implementation of this interface and decide how they want to handle this. I've also written a default implementation of these hooks, which mimics the current behaviour of removing a job from the jobstore. The interface also allows a cloud extension to handle success and failure differently.