Count imported items using the JobStore (#742)
The changes in this pull request are meant to allow the counting of the successfully imported items; The idea is to use a map containing pairs of the form (dataName, dataCount), which is being passed from the current imported DataModel to the ImportResult. Finally, the elements of the map are merged into an accumulator map inside the JobStore.