Fix wrong authurl when doing oauth on devserver
Summary:
Currently, demo server web UI does not redirect to Auth URL to service
providers. Instead it would revert back to localhost:3000 to start
everything over again. This is due to the backend server does not
construct the auth URL correctly.

This commit is to fix that.

Test Plan: Demo UI should proceed to auth url for user to authorize as expected.

Differential Revision: https://phabricator.intern.facebook.com/D21347375