Add new exception to capture user permission denial
Summary:
Currently, when user import photos to google photos, the user needs to
select the google account with access to photos.

Otherwise, Google API would reply with a 403 error which cause a Java
IOException.

This commit is to capture this 403 error in a new exception class based
on CopyExceptionWithFailureReason. So it would allow displaying better
error message to the front end users.

Test Plan:
Import photos to a google account with no access to photos. Worker
should throw a PermissionDeniedException rather than an IOException

Differential Revision: https://phabricator.intern.facebook.com/D21349485