Add finishedJob metric when job is cancelled
Summary:
Currently when a job is started and then cancelled by user, dtp worker
would only print the startedJob metrics line in the log. This makes the
log line incomplete and confuses engineers with unfinished job.

The commit here is to add the finishedJob metrics in job cancellation.

Test Plan: Cancel a job and see the new log line appearing there.

Differential Revision: https://phabricator.intern.facebook.com/D21449063