Fix OOM issue for video task (#895)
Summary:
GoogleVideosImporter is leaking memory because
1) It creates a PhotosLibraryClient every time it imports a video
2) PhotosLibraryClient would create a bunch of worker Gax threads
3) PhotosLibraryClient would not clean up those threads when it finishes

This resuls in OOM issue when transferring large videos jobs.

This commit is to fix this OOM issue by reusing same PhotosLibraryClient
for the same job. So worker is not wasting cpu and memory to create new
threads for each video.

Test Plan:
Transfer many videos in one job and confirm that threads count will not
increase.