Handle NullPointerException when unwrapping Google Photos response
Summary:
Previously we got a null pointer exception when importing a single photo into Google Photos.  This commit adds null checks and instead throws an IOException, which can be used by our workers to retry a photo import.

Test Plan: Ran a test transfer in Docker. Add a test to confirm that an IOException is raised, not a NullPointerException.

Reviewers: #dtp

Differential Revision: https://phabricator.intern.facebook.com/D22459579