Implement alternative stack-based portability data copier
Summary:
Add an alternative implementation of the PortabilityInMemoryDataCopier. This new implementation uses an explicit stack rather than the implicit call stack used in the recursive calls in the original PortabilityInMemoryDataCopier. 

For now this stack is stored locally, however this paves the way to store the stack in external storage.

Test Plan:
Wrote tests to check the functional equivalence of the new PortabilityStackInMemoryDataCopier and the original PortabilityInMemoryDataCopier.

These all pass, as do the previous tests.

Ran a test transfer using the PortabilityStackInMemoryDataCopier.

Reviewers: katriel, wmorland, #dtp

Subscribers: wmorland

Differential Revision: https://phabricator.intern.facebook.com/D22598240

Tasks: T69312083, T69311493