Add option to choose in-memory data copier implementation
Summary:
This diff adds the option to choose the implementation of InMemoryDataCopier, defaulting to the existing behaviour (PortabilityInMemoryDataCopier) if the new PortabilityStackInMemoryDataCopier implementation is not linked in.

The implementation is similar to the IdempotentImportExecutor - it uses a Loader and Extension class. Unlike IdempotentImportExecutor we do not instantiate an implementation, instead we return the class of the implementation. We then bind this in the WorkerModule - this preserves the existing bind behaviour.

Test Plan:
Ran a test transfer, linking in the new PortabilityStackInMemoryDataCopier to the docker demo container. Logs indicate that this class was used in the transfer. Also ran a test transfer without linking in the new implementation - this ran successfully too and chose the PortabilityInMemoryDataCopier as the default implementation.

{F245043890}

Reviewers: #dtp

Differential Revision: https://phabricator.intern.facebook.com/D22647473