Handle case where empty stack loaded from job store
Summary: If the job crashes at the end of the last iteration before updating the job to FINISHED, it would store an empty stack. In this case, we should load the empty stack, and thus avoid further copy iterations. The previous (incorrect) behaviour would restart the job upon seeing an empty stack.

Test Plan: Add test verifying that no further copy iterations are performed if we load an empty job stack.

Reviewers: #dtp

Differential Revision: https://phabricator.intern.facebook.com/D22791600