Update subtypes of ContainerResource
Summary:
When running tests for D22844917, I noticed that VideoContainerResource was not being deserialised correctly as the error message said it was not a subtype of ContainerResource.

This diff corrects the ContainerResource file to add the other subtypes of Container resource, the only exception being BlobbyDataContainerResource - this is because it resides in another module and doing so would introduce a cyclic build dependency. (https://github.com/google/data-transfer-project/issues/916 already flags this). 

Test Plan: Ran ./gradlew build and ./gradlew test

Reviewers: #dtp

Differential Revision: https://phabricator.intern.facebook.com/D22867086