Define DTP Data Model based on Activity Streams
Summary:
AS associates an "actor" with each activity "object". For DTP, we have the same actor for all activity objects, so we disentangle these concepts:
- `SocialActivityActor` corresponds to the "actor"
 - `SocialActivityModel.java` corresponds to the Activity object (`ASObject`).  The name follows other datatypes' conventions e.g. `PhotoModel`.

The `ASObject` contains two fields `image` and `attachments` for images and links respectively. However these fields take in objects with identical fields, so we unify these into a single `attachments` field of type `SocialActivityAttachment` and use a `SocialActivityAttachmentType` enum to distinguish between these cases. This will allow us to allow other types of attachments e.g. videos in future.

The ASObject stored the type of the activity as a string e.g. "check-in", "post" etc. We instead define this more precisely using a  `SocialActivityType` enum.

Finally we define a `SocialActivityLocation` object corresponding to the ASObject's `location` field.

Test Plan:
./gradlew build

Compared data model definition against https://www.w3.org/TR/activitystreams-core/ and the usage of AS in the existing adaptors.

You can see the changes in existing use cases between AS and this implementation in D23564244.

Reviewers: wmorland, #dtp

Reviewed By: wmorland

Differential Revision: https://phabricator.intern.facebook.com/D23564357

Tasks: T71597511

Tags: data-transfer-project

Signature: 23564357:1599484846:4ea0f6ccf070beaa0096574f6205936dea2e3a89