Migrate SocialActivityContainerResource from Activity Streams to use DTP data models
Summary:
We update the `SocialActivityContainerResource` to use the `SocialActivityModel` instead of AS Activity. We store the actor as separate field, so the information is stored per-container-resource rather than the AS per-activity approach.

We also remove `subContainers` property, as it is not relevant to posts (we can't have nested posts within posts). This field has not been used in the GPlusExporter nor the Mastodon importer/exporter.

Test Plan:
./gradlew build

Added tests to check serialisation-deserialisation logic of container.

Reviewers: wmorland, #dtp

Reviewed By: wmorland

Differential Revision: https://phabricator.intern.facebook.com/D23564216

Tasks: T71597511

Tags: data-transfer-project

Signature: 23564216:1599491235:ab80e3ef352479f24f4afdfad39417db6f193855