Use long lived access token for IG photos transfer
Summary:
By default, IG access token is shortlived and only valid for 1 hour.
After that, the graph api query would fail with 400 error.

To fix it normally, we would refresh the token when the short lived
token is expired. However in IG case, there is no refresh token returned
with shortlived access token. So if access token is expired, user would
need to re-authorize again.

Obviously this is not going to work because dtp worker would not be able
to ask user to reauthorize again during a transfer.

So instead, we can exchange the shortlived access token to a long living
one (60 days) as soon as we start the first transfer.

Test Plan:
Used a debugger to check the longlived access token retrievel and tested
the normal transfer on IG photos and that worked.

Reviewers: #dtp

Subscribers:

Tasks:

Tags:

Blame Revision:

Differential Revision: https://phabricator.intern.facebook.com/D23732042