Reduce duplicate logging in CallableImporter (#953)
Summary:
Currently CallableImporter throws out an IOException if
IdempotentImportExecutor recorded errors. This IOExcpetion would be
printed to severe logging if the subsequent retry failed. However, the
content of the exception contains first error and error counts. And
these are already duplicatedly logged in IdempotentImportExecutor. This
cause logs to spam and hard to categorize errors.

This commit is to simplify that by removing the spamming part of the
log.

Test Plan: Ran and checked the log line

Reviewers: #dtp

Subscribers: 

Tasks: T81987523

Tags: 

Blame Revision: 


Differential Revision: https://phabricator.intern.facebook.com/D25995302