Readd log lines for first retryable error
Summary:
We removed the error details from the logline in previous commit.
However, this log line is essential to our retry logic since we are
doing string matching to determine how worker is going to do the retry.

This commit is to add the first error back to the log line.

Test Plan: Build and verify the logline

Reviewers: #dtp

Subscribers: 

Tasks: 

Tags: 

Blame Revision: 


Differential Revision: https://phabricator.intern.facebook.com/D26124640