Only add exception of error details to retryable exception
Summary:
Currently we are logging out both id and title of error details to
Retryable Import Exception. However, this makes the log line hard to be
categorized.

Since we are only logging the first error and we log error for
individual transfer item in the end, we can remove id and title from
here.

This commit is to do that.

Test Plan: build and check log line

Reviewers: #dtp

Subscribers: 

Tasks: T81987523

Tags: 

Blame Revision: 


Differential Revision: https://phabricator.intern.facebook.com/D26147813