Updating Job Watcher service
Changes done:
Updating JobWatching service to check the job state as Preempted or errored.
This is one of the edge case which is to check whether the current job is errored or preempted, and stop the worker job all together.

Earlier, the JobWatching service was exiting the system in case job is cancelled.