Update OAuth1Config for additional parameters (#989)
* Update OAuth1 configs to supply multiple additional parameters

The OAuth1 spec allows parameters on the token request and authorization
URLs. The old configs only supported a single "scope" parameter. This
allows multiple parameters, based on the combination or URL, data type,
and mode.

Adds an additional parameter for SmugMug to allow reading and writing of
private albums as necessary.

* Simplify logic

Co-authored-by: Philip Sharp <philip@smugmug.com>