Backblaze refactor (#1011)
* In this commit I refactor BackblazePhotoImporter to enable unit testing (see next commit)

* Add a test file for BackblazePhotosImporter and update the gradle file with new dependencies

In this commit I added a test file for BackblazePhotosImporter and updated the gradle file with new dependencies required to run it.

* Refactor BackblazeVideosImporter to enable unit testing

Commit is corresponding to the one for the photos importer but this time we're changing the constructor of the video importer. The unit tests are added in the next commit

* Add a test file for  BackblazeVideosImporter

In this commit I'm adding a file with unit tests for the video importer

* Adding BaseBackblazeS3ClientFactory to enable unit testing on the client

The following two commits refacor BackblazeDataTransferClient for facilitating BackblazeDataTransferClient unit testing. In this one I'm adding interface for client factory and adding a base class

* Refactor BackblazeDataTransferClient and BackblazeDataTransferClientFactory to make use of BaseBackblazeS3ClientFactory

In this commit I refactor BackblazeDataTransferClient and BackblazeDataTransferClientFactory to make use of BaseBackblazeS3ClientFactory

* Add BackblazeDataTransferClientTest

In the last commit I implement BackblazeDataTransferClientTest

* Remove unused variable from BackblazeDataTransferClient

Co-authored-by: William Morland <wmorland@fb.com>