noop(cleanup) rm file dead as of c8cce657
`how-does-dtp-work.md` is the original file name of what's now
`documentation.md` and `how-does-dtp-work.md` was just never deleted.
That is: `git mv how-does-dtp-work.md documentation.md` was the
intention of c8cce657 but the old filename was left around and not
linked to by anything, so the file has been dead since then.