noop(cleanup) clarify these are else-branches
this should be noop; adding a tiny bit of doc but more importantly
clarifying that this Photo code we forked is actually composed of what
are _effectively_ else-branches where the code falls through in a "this
should never happen" fashion. Instead we make this explicit by
continuing to the next iteration of our loop.

motivation: Only someone with knowledge of this MicrosoftDriveItem API
would know this, and a reader of this this code won't know that API
necessarily. But the writer of this code certainly knows this, so we
encode that knowledge.