WIP: sets roadmap to complete unit test
comparing to the history of the forked Exporter _itself_ - the core
thing that changed was adding video, but most of the code stayed
identical. this unit test then leaves the core "photo set" unit test in
place and simply adds another for video, trying to mirror the fork of
the exporter itself as closely as possible.