Google media exporter: installs along side Photos, but will replace photos in the future (#1218)
`GoogleMediaExporter` class mosty written by seehamrun@ with later additional commits by jzacsh@

* Create MediaContainerResource to be used by the new MediaExporters and MediaImporters

      This combines the photo and video resources

* Account for videos in the transmogrification logic

* Fix typo in the Transmogrification config

* Add Video limits to transmogrification config

* add clean title to the transmogrification config for videos

* undo rename of the videomodel (let this get handled when the PR to rename gets pushed)

* fix warning abut overriding equals but not hashcode

* remove logic to split up the albums in the media resource

* slight fix to the transmogrification function -- move up the root album to the main function

* Update MediaContainer to use the new VideoModel instead of VideoObject

* Use Album and Video transmogrification specific variables

* Fix return case when maxLength <= 0

* Initial skeleton of the GoogleMediaExporter

This is currently just a copy of the GooglePhotosExporter without all of the content for exporting videos

* Hookup the mediaexporter to the transfer extension

* Add support for video exports in the GoogleMediaExporter

* fix importer

* move convertToVideoModel and convertToPhotoModel into the GoogleMediaItem class

* use getFileName as the title for videos

* Rename PopulateContainedPhotosList method to PopulateContainedMediaList

* Rename `exportPhotos` to `exportMedia`

* Add video support for the exportMediaContainerResource function

jzacsh@'s continued commits:

* fix broken tests: have new data & exceptions

both video and photo models have a new "upload time" field so this
commit _starts_ to introduce that.

also tweaked some of the existing code that was written before
Upload...Exception was added to some signatures.

* nit: consistency in inline-documentation

* nit: stick to existing pattern (of always having a setter)

* noop(docs) leave trail to cleanup since this PR is forking code

* noop(refactor) mv Temp{Photo,Media}Data

noop as this is just a rename of current temp-store logic that is
hardcoded to Photos. Later work will come along to extend the
TempMediaData object to handle new data (added TODO to this point atop
the class).

* WIP: fork of Photos tests trivial "passing"


* noop(refactor) make name match the forked purpose of this token prefix

* WIP: sets roadmap to complete unit test

comparing to the history of the forked Exporter _itself_ - the core
thing that changed was adding video, but most of the code stayed
identical. this unit test then leaves the core "photo set" unit test in
place and simply adds another for video, trying to mirror the fork of
the exporter itself as closely as possible.

* noop(prefactor) clarify photo-logic in tests

this clarifies what the photo-specific fakes are between a unit test's
arrange and assert phases. that's not always clear for _other_ consts
that are pulled to the top of the test file, but I'm trying to merely
unblock the addition of video logic here (not do a _general_ cleanup).

* bugfix: all the fakes should be readonly

* finishes unit test by fixing URL assertion

while I'm at it, also refactoring the Media objects involved so google
adapter logic isn't doing deep property-null-checking as an API (`isVideo`
`isPhoto`).

---------

Co-authored-by: Siham Hussein <sihamh@google.com>