WIP: refork Photos importer + factor out Video logic
re-forking Photos code (as that's how this branch was originally
developed) because a lot of upstream changes have landed in video code
since then.

while I'm at it I'm also doing what I did with
MicrosoftMedia{Importer,Exporter} and trying to make the MediaImporter
not repeat itself, but at least factor our the common logic looping over
video vs. looping over photos objects internally.

this refactoring isn't perfect, but it's a baby step: in the future we
can use the abstractions added internally to this class to let _all_ the
adapters in the google package reuse this same code (both the Photos and
the Videos adapters).

state of this commit: doesn't compile quite yet:

  ```
  $ gradle test
  > Task :extensions:data-transfer:portability-data-transfer-google:compileJava FAILED
  /usr/local/google/home/zacsh/pub/src/dtp/extensions/data-transfer/portability-data-transfer-google/src/main/java/org/datatransferproject/datatransfer/google/media/GoogleMediaImporter.java:188: error: incompatible types: invalid method reference
          this::importPhotoBatch);
          ^
      incompatible types: List<T> cannot be converted to List<PhotoModel>
    where T is a type-variable:
      T extends Object declared in method <T>importToAlbum(UUID,TokensAndUrlAuthData,List<T>,IdempotentImportExecutor,String)
  Note: /usr/local/google/home/zacsh/pub/src/dtp/extensions/data-transfer/portability-data-transfer-google/src/main/java/org/datatransferproject/datatransfer/google/videos/GoogleVideosInterface.java uses or overrides a deprecated API.
  Note: Recompile with -Xlint:deprecation for details.
  Note: Some input files use unchecked or unsafe operations.
  Note: Recompile with -Xlint:unchecked for details.
  Note: Some messages have been simplified; recompile with -Xdiags:verbose to get full output
  1 error

  FAILURE: Build failed with an exception.
  ```