noop(dev) make gradle cli (`gradle test`) friendly
before this commit the only way to see the reason for an assertion in a
unit test failure was to open the gradle-generated HTML file in a
browser (not always convenient, eg: if you're developing via SSH,
headless).

now the particular reason for a unit test failure is also printed to
stderr as you'd expect when running tests on a commandline.

Revert "prefactor[squash me]: simplify by reverting `gradle test` tweak"

This reverts commit 9de0e7d07d97959bd156274b3d1fb31e4fef69f6.