resolve TODO: match full stacktrace retry strategy
this is a new feature that should have no impact on existing
RetryStrategy yaml/json users: there's simply a new `stacktraceRegexes`
key in the config that acts as an alternative to the older `regexes`.
It's only difference from the original is that its regexes match across
the long, multi-line strings that are **full** stacktraces.

motivation: matching more of the Throwable is a TODO that's come up
multiple times and now actually caused a surprise last-minute bug for
us, so this patch resolves the feature gap and unblocks google's and
apple's work.

testing: for now we're leaving backfilling (a new unit test for this
file) as a fast-follow, since Google has intern integration~esque
unit-test coverage of this file against our real-world YAML config. Not
ideal, but need time to come back and opensource the test.