fix microsoft OOOM: don't stream into memory
I'm fixing a bug here where the implementation of the microsoft adapter
was reading an arbitrary stream into memory (using DataChunk class), and
instead "just stream directly to the endpoint". It's not _quite_ that
simple though: we actually stream twice, for the very reason that the
old implementation was streaming into memory: we need to know the
filesize.

while we're here I'm doing some "leave the campground better" tasks, so
here's a more nitty-gritty breakdown of this PR:
-  microsoft adapter: marking duplicated code Microsoft{Photo,Video}* as
   needing to rely on newly refactored MicrosoftMedia (so hese kinds of
   bug fixes are easier to maintain, for example).
- DTP: make testability (via DI) easier with for java.net.URL streamers
  (this has already become a pattern, so I just formalized it and
  dropped a TODO in the places that are doing this locally; this way, in
  the future it'll be easier/more obvious what "the pattern" is and how
  to stop maintaining disparate copies of it)
- microsoft adapter: remainder of DataChunk code is now just a POJO, so
  switching to autovalue and letting the test live in the primary user:
  the new `StreamChunker` (as a mini "integrated" test).
- microsoft adapter: all size-related operations are a `long` now