noop(cleanup) run clang-format (enforce goog style)
I have to fight this everytime I'm working on or fixing something in
this adapter. so just let it run once and for all and merge as a
standalone PR so it's clearly separated from any other changes one might
be trying to debug.

for more, see the extant DTP documentation on developer setup:
https://github.com/dtinit/data-transfer-project/blob/badda9bb520e59cd13227e2f4a9e4189b0bc7e59/Documentation/Developer.md#setup-formatting
(I'm not using intellij, but google-java-format exists in every editor
I've used, so the instructions should still stand).