runtime-bugfix: reading value always being closed (#1410)
while we're at it, just take what we need (unmarshal into a string) and close it
ourselves, letting caller stop worrrying about this closeable.

while we're he we *also* cleanup all our unit tests so they use real Response
objects, and stop mocking it. (This is the same recent codehealth change made to
stop using ResponseBody mocks, and use the real things). Not interacting with
the real OkHTTP API was part of what made this issue invisible before.