ci: release without committing to main branch
Committing to the master branch isn't allowed from github actions
because of our branch protection rules. There's no easy way to allow
github actions to bypass this.

As a workaround, skip all the parts of the release that needed to commit
to the repo. This means moving the version config to an environment
variable, which shouldn't be an issue.