fix(generic-adapters): make ContainerSerializer interface public
The `ContainerSerializer` interface needs to be public for any package
other than `org.datatranfserproject.datatransfer.generic` to be able to
instantiate a `GenericImporter`.
We want to implement a custom `GenericTransferExtension` that can load
from a database rather than packaged config file, so need to be able to
instantiate `GenericImporter`s from a different package.