fix: googphotos http request handling to be self-consistent (#1445)
factors out the common HTTP logic to a middle-tier private func
`makeHttpRequest` to get to a _mostly_ noop cleanup/DRYing of our HTTP
handling. (motivation: some issues we're debugging). Things that prevent
this from being just a true **noop** refactor:

- `makePostRequest` used to only respect its users' request for special
HTTP headers on the _first_ request, but not on subsequent
(token-refresh driven) requests. Now it'll respect it for all requests.
- similar to previous point: _rate limiting_ wasn't consistently
applied, now it's always applied for every write request, and not just
the first.
- similarly, check for HTTP 200 was only done on some request paths, now
it's always done; it's also now a consistent stack trace that will
hopefully be more useful (like the other stack traces we get for bad
responses).
- unlike the above/extant http-200 check (despite not getting an error),
other checks like permission-denied were not always done; now they're
done regardless of getting a response or an error